backend/requirements.txt
# 1. django 5.0.1이상 6.0.0 미만
django>=5.0.1,<6.0.0
# 2. django rest framework
djangorestframework>=3.14.0,<4.0.0
# 3. django rest framework simplejwt auth
djangorestframework-simplejwt>=5.3.1,<6.0.0
# 4. mysql - mysql adapter
mysqlclient>=2.2.4,<3.0.0
# 5. api docs
drf-spectacular>=0.27.1,<0.28.0
# 6.uWSGI 
uWSGI>=2.0.24,<2.1.0
# 7. django-crontab 스케줄에 따른 실행 라이브러리
django-crontab>=0.7.1,<0.8.0
# 8. dotenv 로컬환경에서 실행시키기위함
python-dotenv>=1.0.1,<2.0.0
# 9. django-cors-headers : cors 이슈 해결
django-cors-headers>=4.3.1,<5.0.0
# 10. django-allauth
django-allauth>=0.61.1,<1.0.0